fix(test): update phpspec to 6.0 version
Problem
-------

phpspec was set in version ^4. This version was not compatible with
PHP 7.4. So if you have the latest version of PHP on your computer
you
cannot develop on this plugin.

And the upgrade was blocking because of another dependency.

Solution
--------

Update phpspec to ^6 version. Change phpspec-code-coverage to the
friends-of-phpspec version which allows us to use phpspec ^6.